popularity-contest/requirements.txt
functions-framework==3.*
google-cloud-storage
